Move basic search bits and bobs into their own module
